Fixed input and output directory manipulation.
If one forget to add a trailing slash in the input or output global variables
then filenames were build incorrectly. The slash between the input (output)
part and the rest of the page source file was forgotten resulting in an
inexistent file path.

I add some code to ensure that both variables (input and output) embed a
unique trailing slash when initializing the whole process (in Site::init).