Propagate the --debug flag only if it has been found.
Creating the Templer::Global object with debug key fixed to the `undef`
value (which is the case if no --debug was set on the command line) cause
troubles. In this case, since debug has a value it cannot be merged with the
default settings and any use of debug field throw an error.

In order to keep the default settings at one place only in the code I added a
test to propagate the --debug to the creator method only if it has been asked.