Update Templer::Site to fix copyAsset behavior (shell safe filenames in tar).
As the copy is done through a shell command filenames should be modified in
order to protect characters with meaning for the shell (space, quote, etc.)

To achieve that I quote-protect every filename when they do not embed quote,
and double-quote-protect them when it does. In that latter case I also need to
protect backslash, dollar and backquote (which keep their meaning in
double-quote string) as well as double-quote itself.