Fixed FileGlob matching files type detection.
The test to detect if a file is an input page was not working (when using
file_glob I got a lot of warning every file which were not images were treated
as input file and never as *regular* ones). I think it was due to the way the
regex was constructed but still cannot understand why.

I also modified tests in order to try to cover much more documented features.

As a side effect, something is still disturbing: one can create a
Templer::Site object with no suffix set up (as it was the case in the tests).