Removed the Macros plugin which appears because of my mistake.
When Steve gave me the push right on his repository I merged the wrong branch
in its master one.

As previously said this plugin is not ready to be included in the main
repository: not written in the templer philosophy (macros should be defined in
header only and in a different, more simple, way than today) and no tests are
available.